return to store large factors of congruences as 64 bit numbers in order to have more options for very large N

+++ src/de/tilman_neumann/jml/factor/base/congruence/Partial.java
@@ -37,5 +37,5 @@ abstract public class Partial extends AQPair {
 	/**
 	 * @return the large factors appearing with odd exponent.
 	 */
-	abstract public Integer[] getLargeFactorsWithOddExponent(); // TODO use int[] ?
+	abstract public Long[] getLargeFactorsWithOddExponent();
 }
